refactor append/prepend to remove duplication
