+++ src/periodicnotes/sectionposition.ts
@@ -1,0 +1,6 @@
+export const SectionPosition = {
+	PREPEND: "prepend",
+	APPEND: "append",
+};
+export type SectionPosition =
+	typeof SectionPosition[keyof typeof SectionPosition];
